Refactor settings editor and widget modules
