+++ src/settings_editor.rs
@@ -1,1001 +1,13 @@
-use crate::dashboard::config::DashboardConfig;
-use crate::gui::{LauncherApp, UiErrorEvent};
-use crate::hotkey::parse_hotkey;
-use crate::plugins::note::{NoteExternalOpen, NotePluginSettings};
-use crate::plugins::screenshot::ScreenshotPluginSettings;
-use crate::settings::{QueryResultsLayoutSettings, Settings};
-use eframe::egui;
-use egui_toast::{Toast, ToastKind, ToastOptions};
-use std::sync::Arc;
+#[path = "settings_editor/mapping.rs"]
+mod mapping;
+#[path = "settings_editor/render.rs"]
+mod render;
+#[path = "settings_editor/save.rs"]
+mod save;
+#[path = "settings_editor/state.rs"]
+mod state;
 
-#[derive(Default)]
-pub struct SettingsEditor {
-    hotkey: String,
-    hotkey_valid: bool,
-    last_valid_hotkey: String,
-    quit_hotkey_enabled: bool,
-    quit_hotkey: String,
-    quit_hotkey_valid: bool,
-    last_valid_quit_hotkey: String,
-    help_hotkey_enabled: bool,
-    help_hotkey: String,
-    help_hotkey_valid: bool,
-    last_valid_help_hotkey: String,
-    debug_logging: bool,
-    show_toasts: bool,
-    show_inline_errors: bool,
-    show_error_toasts: bool,
-    toast_duration: f32,
-    offscreen_x: i32,
-    offscreen_y: i32,
-    window_w: i32,
-    window_h: i32,
-    note_panel_w: f32,
-    note_panel_h: f32,
-    note_save_on_close: bool,
-    note_always_overwrite: bool,
-    note_images_as_links: bool,
-    note_show_details: bool,
-    note_more_limit: usize,
-    query_scale: f32,
-    list_scale: f32,
-    history_limit: usize,
-    clipboard_limit: usize,
-    fuzzy_weight: f32,
-    usage_weight: f32,
-    match_exact: bool,
-    page_jump: usize,
-    query_results_layout_enabled: bool,
-    query_results_layout_rows: usize,
-    query_results_layout_cols: usize,
-    query_results_layout_respect_plugin_capability: bool,
-    query_results_layout_plugin_opt_out: String,
-    follow_mouse: bool,
-    static_enabled: bool,
-    static_x: i32,
-    static_y: i32,
-    static_w: i32,
-    static_h: i32,
-    hide_after_run: bool,
-    pub always_on_top: bool,
-    timer_refresh: f32,
-    disable_timer_updates: bool,
-    preserve_command: bool,
-    clear_query_after_run: bool,
-    require_confirm_destructive: bool,
-    query_autocomplete: bool,
-    net_refresh: f32,
-    net_unit: crate::settings::NetUnit,
-    screenshot_dir: String,
-    screenshot_save_file: bool,
-    screenshot_auto_save: bool,
-    screenshot_use_editor: bool,
-    reduce_dashboard_work_when_unfocused: bool,
-    show_dashboard_diagnostics: bool,
-    dashboard_enabled: bool,
-    dashboard_path: String,
-    dashboard_default_location: String,
-    dashboard_show_when_empty: bool,
-    plugin_settings: std::collections::HashMap<String, serde_json::Value>,
-    plugins_expanded: bool,
-    expand_request: Option<bool>,
-}
-
-impl SettingsEditor {
-    const SETTINGS_WINDOW_DEFAULT_WIDTH: f32 = 640.0;
-    const SETTINGS_WINDOW_MAX_DEFAULT_HEIGHT: f32 = 720.0;
-    const SETTINGS_WINDOW_MIN_HEIGHT: f32 = 360.0;
-    const SETTINGS_CONTENT_MIN_HEIGHT: f32 = 180.0;
-    const SETTINGS_FOOTER_RESERVED_HEIGHT: f32 = 56.0;
-
-    fn normalized_static_settings(
-        follow_mouse: bool,
-        static_enabled: bool,
-        static_pos: (i32, i32),
-        static_size: (i32, i32),
-    ) -> (bool, Option<(i32, i32)>, Option<(i32, i32)>) {
-        if follow_mouse {
-            (false, None, None)
-        } else {
-            (static_enabled, Some(static_pos), Some(static_size))
-        }
-    }
-
-    pub fn from_settings(settings: &Settings) -> Self {
-        let hotkey = settings.hotkey.clone().unwrap_or_default();
-        let hotkey_valid = parse_hotkey(&hotkey).is_some();
-        let default_hotkey = Settings::default().hotkey.unwrap_or_else(|| "F2".into());
-        let last_valid_hotkey = if hotkey_valid {
-            hotkey.clone()
-        } else {
-            default_hotkey.clone()
-        };
-        let quit_hotkey = settings.quit_hotkey.clone().unwrap_or_default();
-        let quit_hotkey_enabled = settings.quit_hotkey.is_some();
-        let quit_hotkey_valid = if quit_hotkey_enabled {
-            parse_hotkey(&quit_hotkey).is_some()
-        } else {
-            true
-        };
-        let last_valid_quit_hotkey = if quit_hotkey_valid {
-            quit_hotkey.clone()
-        } else {
-            String::new()
-        };
-        let help_hotkey = settings.help_hotkey.clone().unwrap_or_default();
-        let help_hotkey_enabled = settings.help_hotkey.is_some();
-        let help_hotkey_valid = if help_hotkey_enabled {
-            parse_hotkey(&help_hotkey).is_some()
-        } else {
-            true
-        };
-        let last_valid_help_hotkey = if help_hotkey_valid {
-            help_hotkey.clone()
-        } else {
-            String::new()
-        };
-        let follow_mouse = settings.follow_mouse;
-        let (static_enabled, _, _) = Self::normalized_static_settings(
-            follow_mouse,
-            settings.static_location_enabled,
-            settings.static_pos.unwrap_or((0, 0)),
-            settings.static_size.unwrap_or((400, 220)),
-        );
-        let mut s = Self {
-            hotkey,
-            hotkey_valid,
-            last_valid_hotkey,
-            quit_hotkey_enabled,
-            quit_hotkey,
-            quit_hotkey_valid,
-            last_valid_quit_hotkey,
-            help_hotkey_enabled,
-            help_hotkey,
-            help_hotkey_valid,
-            last_valid_help_hotkey,
-            debug_logging: settings.debug_logging,
-            show_toasts: settings.enable_toasts,
-            show_inline_errors: settings.show_inline_errors,
-            show_error_toasts: settings.show_error_toasts,
-            toast_duration: settings.toast_duration,
-            offscreen_x: settings.offscreen_pos.unwrap_or((2000, 2000)).0,
-            offscreen_y: settings.offscreen_pos.unwrap_or((2000, 2000)).1,
-            window_w: settings.window_size.unwrap_or((400, 220)).0,
-            window_h: settings.window_size.unwrap_or((400, 220)).1,
-            note_panel_w: settings.note_panel_default_size.0,
-            note_panel_h: settings.note_panel_default_size.1,
-            note_save_on_close: settings.note_save_on_close,
-            note_always_overwrite: settings.note_always_overwrite,
-            note_images_as_links: settings.note_images_as_links,
-            note_show_details: settings.note_show_details,
-            note_more_limit: settings.note_more_limit,
-            query_scale: settings.query_scale.unwrap_or(1.0),
-            list_scale: settings.list_scale.unwrap_or(1.0),
-            history_limit: settings.history_limit,
-            clipboard_limit: settings.clipboard_limit,
-            fuzzy_weight: settings.fuzzy_weight,
-            usage_weight: settings.usage_weight,
-            match_exact: settings.match_exact,
-            page_jump: settings.page_jump,
-            query_results_layout_enabled: settings.query_results_layout.enabled,
-            query_results_layout_rows: settings.query_results_layout.rows.max(1),
-            query_results_layout_cols: settings.query_results_layout.cols.max(1),
-            query_results_layout_respect_plugin_capability: settings
-                .query_results_layout
-                .respect_plugin_capability,
-            query_results_layout_plugin_opt_out: settings
-                .query_results_layout
-                .plugin_opt_out
-                .join(", "),
-            follow_mouse,
-            static_enabled,
-            static_x: settings.static_pos.unwrap_or((0, 0)).0,
-            static_y: settings.static_pos.unwrap_or((0, 0)).1,
-            static_w: settings.static_size.unwrap_or((400, 220)).0,
-            static_h: settings.static_size.unwrap_or((400, 220)).1,
-            hide_after_run: settings.hide_after_run,
-            always_on_top: settings.always_on_top,
-            timer_refresh: settings.timer_refresh,
-            disable_timer_updates: settings.disable_timer_updates,
-            preserve_command: settings.preserve_command,
-            clear_query_after_run: settings.clear_query_after_run,
-            require_confirm_destructive: settings.require_confirm_destructive,
-            query_autocomplete: settings.query_autocomplete,
-            net_refresh: settings.net_refresh,
-            net_unit: settings.net_unit,
-            screenshot_dir: settings.screenshot_dir.clone().unwrap_or_default(),
-            screenshot_save_file: settings.screenshot_save_file,
-            screenshot_auto_save: settings.screenshot_auto_save,
-            screenshot_use_editor: settings.screenshot_use_editor,
-            reduce_dashboard_work_when_unfocused: settings.reduce_dashboard_work_when_unfocused,
-            show_dashboard_diagnostics: settings.show_dashboard_diagnostics,
-            dashboard_enabled: settings.dashboard.enabled,
-            dashboard_path: settings
-                .dashboard
-                .config_path
-                .clone()
-                .unwrap_or_else(|| "dashboard.json".into()),
-            dashboard_default_location: settings
-                .dashboard
-                .default_location
-                .clone()
-                .unwrap_or_default(),
-            dashboard_show_when_empty: settings.dashboard.show_when_query_empty,
-            plugin_settings: settings.plugin_settings.clone(),
-            plugins_expanded: false,
-            expand_request: None,
-        };
-        s.plugin_settings
-            .entry("screenshot".into())
-            .or_insert_with(|| {
-                serde_json::json!({
-                    "screenshot_dir": s.screenshot_dir.clone(),
-                    "screenshot_save_file": s.screenshot_save_file,
-                    "screenshot_auto_save": s.screenshot_auto_save,
-                    "screenshot_use_editor": s.screenshot_use_editor,
-                })
-            });
-        s
-    }
-
-    pub fn new(settings: &Settings) -> Self {
-        Self::from_settings(settings)
-    }
-
-    pub fn new_with_plugins(settings: &Settings) -> Self {
-        let mut s = Self::new(settings);
-        s.sync_from_plugin_settings();
-        s
-    }
-
-    fn settings_window_default_height(available_height: f32) -> f32 {
-        (available_height * 0.5)
-            .max(Self::SETTINGS_WINDOW_MIN_HEIGHT)
-            .min(Self::SETTINGS_WINDOW_MAX_DEFAULT_HEIGHT)
-    }
-
-    fn settings_window_default_size(ctx: &egui::Context) -> [f32; 2] {
-        [
-            Self::SETTINGS_WINDOW_DEFAULT_WIDTH,
-            Self::settings_window_default_height(ctx.available_rect().height()),
-        ]
-    }
-
-    fn settings_content_height(available_height: f32) -> f32 {
-        (available_height - Self::SETTINGS_FOOTER_RESERVED_HEIGHT)
-            .max(Self::SETTINGS_CONTENT_MIN_HEIGHT)
-    }
-
-    fn sync_from_plugin_settings(&mut self) {
-        if let Some(val) = self.plugin_settings.get("clipboard") {
-            if let Ok(cfg) = serde_json::from_value::<
-                crate::plugins::clipboard::ClipboardPluginSettings,
-            >(val.clone())
-            {
-                self.clipboard_limit = cfg.max_entries;
-            }
-        }
-        if let Some(val) = self.plugin_settings.get("network") {
-            if let Ok(cfg) = serde_json::from_value::<crate::plugins::network::NetworkPluginSettings>(
-                val.clone(),
-            ) {
-                self.net_refresh = cfg.refresh_rate;
-                self.net_unit = cfg.unit;
-            }
-        }
-        if let Some(val) = self.plugin_settings.get("history") {
-            if let Ok(cfg) = serde_json::from_value::<crate::plugins::history::HistoryPluginSettings>(
-                val.clone(),
-            ) {
-                self.history_limit = cfg.max_entries;
-            }
-        }
-        if let Some(val) = self.plugin_settings.get("screenshot") {
-            if let Ok(cfg) = serde_json::from_value::<ScreenshotPluginSettings>(val.clone()) {
-                self.screenshot_dir = cfg.screenshot_dir;
-                self.screenshot_save_file = cfg.screenshot_save_file;
-                self.screenshot_auto_save = cfg.screenshot_auto_save;
-                self.screenshot_use_editor = cfg.screenshot_use_editor;
-            }
-        }
-    }
-
-    /// Returns the raw JSON value for a plugin's settings, if present.
-    pub fn get_plugin_setting_value(&self, name: &str) -> Option<&serde_json::Value> {
-        self.plugin_settings.get(name)
-    }
-
-    /// Sets the raw JSON value for a plugin's settings and re-syncs any derived editor state.
-    pub fn set_plugin_setting_value(&mut self, name: &str, value: serde_json::Value) {
-        self.plugin_settings.insert(name.to_string(), value);
-        self.sync_from_plugin_settings();
-    }
-
-    pub fn to_settings(&self, current: &Settings) -> Settings {
-        let (static_location_enabled, static_pos, static_size) = Self::normalized_static_settings(
-            self.follow_mouse,
-            self.static_enabled,
-            (self.static_x, self.static_y),
-            (self.static_w, self.static_h),
-        );
-        Settings {
-            hotkey: if self.hotkey.trim().is_empty() {
-                None
-            } else {
-                Some(self.hotkey.clone())
-            },
-            quit_hotkey: if !self.quit_hotkey_enabled || self.quit_hotkey.trim().is_empty() {
-                None
-            } else {
-                Some(self.quit_hotkey.clone())
-            },
-            help_hotkey: if !self.help_hotkey_enabled || self.help_hotkey.trim().is_empty() {
-                None
-            } else {
-                Some(self.help_hotkey.clone())
-            },
-            index_paths: current.index_paths.clone(),
-            max_indexed_items: current.max_indexed_items,
-            plugin_dirs: current.plugin_dirs.clone(),
-            enabled_plugins: current.enabled_plugins.clone(),
-            enabled_capabilities: current.enabled_capabilities.clone(),
-            debug_logging: self.debug_logging,
-            log_file: current.log_file.clone(),
-            enable_toasts: self.show_toasts,
-            show_inline_errors: self.show_inline_errors,
-            show_error_toasts: self.show_error_toasts,
-            toast_duration: self.toast_duration,
-            offscreen_pos: Some((self.offscreen_x, self.offscreen_y)),
-            window_size: Some((self.window_w, self.window_h)),
-            note_panel_default_size: (self.note_panel_w, self.note_panel_h),
-            note_save_on_close: self.note_save_on_close,
-            note_always_overwrite: self.note_always_overwrite,
-            note_images_as_links: self.note_images_as_links,
-            note_show_details: self.note_show_details,
-            note_more_limit: self.note_more_limit,
-            query_scale: Some(self.query_scale),
-            list_scale: Some(self.list_scale),
-            history_limit: self.history_limit,
-            clipboard_limit: self.clipboard_limit,
-            fuzzy_weight: self.fuzzy_weight,
-            usage_weight: self.usage_weight,
-            match_exact: self.match_exact,
-            page_jump: self.page_jump,
-            query_results_layout: QueryResultsLayoutSettings {
-                enabled: self.query_results_layout_enabled,
-                rows: self.query_results_layout_rows.max(1),
-                cols: self.query_results_layout_cols.max(1),
-                respect_plugin_capability: self.query_results_layout_respect_plugin_capability,
-                plugin_opt_out: self
-                    .query_results_layout_plugin_opt_out
-                    .split(',')
-                    .map(str::trim)
-                    .filter(|s| !s.is_empty())
-                    .map(ToOwned::to_owned)
-                    .collect(),
-            },
-            follow_mouse: self.follow_mouse,
-            static_location_enabled,
-            static_pos,
-            static_size,
-            hide_after_run: self.hide_after_run,
-            always_on_top: self.always_on_top,
-            timer_refresh: self.timer_refresh,
-            disable_timer_updates: self.disable_timer_updates,
-            preserve_command: self.preserve_command,
-            clear_query_after_run: self.clear_query_after_run,
-            require_confirm_destructive: self.require_confirm_destructive,
-            query_autocomplete: self.query_autocomplete,
-            net_refresh: self.net_refresh,
-            net_unit: self.net_unit,
-            screenshot_dir: if self.screenshot_dir.trim().is_empty() {
-                None
-            } else {
-                Some(self.screenshot_dir.clone())
-            },
-            screenshot_save_file: self.screenshot_save_file,
-            screenshot_auto_save: self.screenshot_auto_save,
-            screenshot_use_editor: self.screenshot_use_editor,
-            plugin_settings: self.plugin_settings.clone(),
-            show_examples: current.show_examples,
-            pinned_panels: current.pinned_panels.clone(),
-            reduce_dashboard_work_when_unfocused: self.reduce_dashboard_work_when_unfocused,
-            show_dashboard_diagnostics: self.show_dashboard_diagnostics,
-            dashboard: crate::settings::DashboardSettings {
-                enabled: self.dashboard_enabled,
-                config_path: if self.dashboard_path.trim().is_empty() {
-                    None
-                } else {
-                    Some(self.dashboard_path.clone())
-                },
-                default_location: if self.dashboard_default_location.trim().is_empty() {
-                    None
-                } else {
-                    Some(self.dashboard_default_location.clone())
-                },
-                show_when_query_empty: self.dashboard_show_when_empty,
-            },
-            theme: current.theme.clone(),
-            note_graph: current.note_graph.clone(),
-        }
-    }
-
-    pub fn ui(&mut self, ctx: &egui::Context, app: &mut LauncherApp) {
-        let mut open = app.show_settings;
-        egui::Window::new("Settings")
-            .resizable(true)
-            .default_size(Self::settings_window_default_size(ctx))
-            .min_height(Self::SETTINGS_WINDOW_MIN_HEIGHT)
-            .open(&mut open)
-            .show(ctx, |ui| {
-                let settings_content_height = Self::settings_content_height(ui.available_height());
-                egui::ScrollArea::vertical()
-                    .max_height(settings_content_height)
-                    .show(ui, |ui| {
-                        ui.horizontal(|ui| {
-                            ui.label("Launcher hotkey");
-                            let resp = ui.text_edit_singleline(&mut self.hotkey);
-                            if resp.changed() {
-                                self.hotkey_valid = parse_hotkey(&self.hotkey).is_some();
-                                if self.hotkey_valid {
-                                    self.last_valid_hotkey = self.hotkey.clone();
-                                }
-                            }
-                            let color = if self.hotkey_valid {
-                                egui::Color32::GREEN
-                            } else {
-                                egui::Color32::RED
-                            };
-                            ui.add(egui::Label::new(egui::RichText::new("●").color(color)));
-                        });
-                        ui.checkbox(&mut self.quit_hotkey_enabled, "Enable quit hotkey");
-                        if self.quit_hotkey_enabled {
-                            ui.horizontal(|ui| {
-                                ui.label("Quit hotkey");
-                                let resp = ui.text_edit_singleline(&mut self.quit_hotkey);
-                                if resp.changed() {
-                                    self.quit_hotkey_valid =
-                                        parse_hotkey(&self.quit_hotkey).is_some();
-                                    if self.quit_hotkey_valid {
-                                        self.last_valid_quit_hotkey = self.quit_hotkey.clone();
-                                    }
-                                }
-                                let color = if self.quit_hotkey_valid {
-                                    egui::Color32::GREEN
-                                } else {
-                                    egui::Color32::RED
-                                };
-                                ui.add(egui::Label::new(egui::RichText::new("●").color(color)));
-                            });
-                        }
-
-                        ui.checkbox(&mut self.help_hotkey_enabled, "Enable help hotkey");
-                        if self.help_hotkey_enabled {
-                            ui.horizontal(|ui| {
-                                ui.label("Help hotkey");
-                                let resp = ui.text_edit_singleline(&mut self.help_hotkey);
-                                if resp.changed() {
-                                    self.help_hotkey_valid =
-                                        parse_hotkey(&self.help_hotkey).is_some();
-                                    if self.help_hotkey_valid {
-                                        self.last_valid_help_hotkey = self.help_hotkey.clone();
-                                    }
-                                }
-                                let color = if self.help_hotkey_valid {
-                                    egui::Color32::GREEN
-                                } else {
-                                    egui::Color32::RED
-                                };
-                                ui.add(egui::Label::new(egui::RichText::new("●").color(color)));
-                            });
-                        }
-
-                        ui.horizontal(|ui| {
-                            egui::ComboBox::from_label("Debug logging")
-                                .selected_text(if self.debug_logging {
-                                    "Enabled"
-                                } else {
-                                    "Disabled"
-                                })
-                                .show_ui(ui, |ui| {
-                                    ui.selectable_value(&mut self.debug_logging, false, "Disabled");
-                                    ui.selectable_value(&mut self.debug_logging, true, "Enabled");
-                                });
-                        });
-
-                        ui.checkbox(&mut self.show_toasts, "Enable toast notifications");
-                        ui.checkbox(&mut self.show_inline_errors, "Show inline errors");
-                        ui.checkbox(&mut self.show_error_toasts, "Show error toasts");
-                        if self.show_toasts {
-                            ui.horizontal(|ui| {
-                                ui.label("Toast duration (s)");
-                                ui.add(
-                                    egui::Slider::new(&mut self.toast_duration, 0.1..=5.0).text(""),
-                                );
-                            });
-                        }
-                        ui.horizontal_wrapped(|ui| {
-                            ui.checkbox(
-                                &mut self.hide_after_run,
-                                "Hide window after running action",
-                            );
-                            ui.checkbox(&mut self.preserve_command, "Preserve command after run");
-                            ui.checkbox(&mut self.clear_query_after_run, "Clear query after run");
-                            ui.checkbox(
-                                &mut self.require_confirm_destructive,
-                                "Require confirm for destructive actions",
-                            );
-                        });
-                        ui.checkbox(&mut self.always_on_top, "Always on top");
-                        ui.checkbox(&mut self.query_autocomplete, "Enable query autocomplete");
-                        ui.checkbox(
-                            &mut self.disable_timer_updates,
-                            "Disable timer auto refresh",
-                        );
-                        ui.horizontal(|ui| {
-                            ui.label("Timer refresh rate (s)");
-                            ui.add_enabled_ui(!self.disable_timer_updates, |ui| {
-                                ui.add(
-                                    egui::DragValue::new(&mut self.timer_refresh)
-                                        .clamp_range(0.1..=60.0)
-                                        .speed(0.1),
-                                );
-                            });
-                        });
-
-                        ui.horizontal(|ui| {
-                            ui.label("Query scale");
-                            ui.add(egui::Slider::new(&mut self.query_scale, 0.5..=5.0).text(""));
-                        });
-                        ui.horizontal(|ui| {
-                            ui.label("List scale");
-                            ui.add(egui::Slider::new(&mut self.list_scale, 0.5..=5.0).text(""));
-                        });
-                        if ui.button("Open Theme Settings...").clicked() {
-                            app.open_theme_settings_dialog();
-                        }
-                        ui.horizontal(|ui| {
-                            ui.label("Fuzzy weight");
-                            ui.add(egui::Slider::new(&mut self.fuzzy_weight, 0.0..=5.0).text(""));
-                        });
-                        ui.horizontal(|ui| {
-                            ui.label("Usage weight");
-                            ui.add(egui::Slider::new(&mut self.usage_weight, 0.0..=5.0).text(""));
-                        });
-                        ui.checkbox(&mut self.match_exact, "Match exact");
-
-                        ui.horizontal(|ui| {
-                            ui.label("Page jump");
-                            ui.add(
-                                egui::DragValue::new(&mut self.page_jump)
-                                    .clamp_range(1..=100)
-                                    .speed(1),
-                            );
-                        });
-
-                        ui.checkbox(
-                            &mut self.query_results_layout_enabled,
-                            "Display results in grid layout",
-                        );
-                        ui.add_enabled_ui(self.query_results_layout_enabled, |ui| {
-                            ui.horizontal(|ui| {
-                                ui.label("Grid rows");
-                                ui.add(
-                                    egui::DragValue::new(&mut self.query_results_layout_rows)
-                                        .clamp_range(1..=100)
-                                        .speed(1),
-                                );
-                                ui.label("Columns");
-                                ui.add(
-                                    egui::DragValue::new(&mut self.query_results_layout_cols)
-                                        .clamp_range(1..=100)
-                                        .speed(1),
-                                );
-                            });
-                            self.query_results_layout_rows = self.query_results_layout_rows.max(1);
-                            self.query_results_layout_cols = self.query_results_layout_cols.max(1);
-                            ui.checkbox(
-                                &mut self.query_results_layout_respect_plugin_capability,
-                                "Respect plugin list/grid capability",
-                            );
-                            ui.horizontal(|ui| {
-                                ui.label("Force list for plugins (comma separated)");
-                                ui.text_edit_singleline(
-                                    &mut self.query_results_layout_plugin_opt_out,
-                                );
-                            });
-                        });
-
-                        ui.horizontal(|ui| {
-                            ui.label("Off-screen X");
-                            ui.add(egui::DragValue::new(&mut self.offscreen_x));
-                            ui.label("Y");
-                            ui.add(egui::DragValue::new(&mut self.offscreen_y));
-                        });
-
-                        ui.horizontal(|ui| {
-                            ui.label("Note panel W");
-                            ui.add(egui::DragValue::new(&mut self.note_panel_w));
-                            ui.label("H");
-                            ui.add(egui::DragValue::new(&mut self.note_panel_h));
-                        });
-
-                        let follow_mouse_resp = ui.checkbox(&mut self.follow_mouse, "Follow mouse");
-                        if follow_mouse_resp.changed() && self.follow_mouse {
-                            self.static_enabled = false;
-                        }
-                        ui.add_enabled_ui(!self.follow_mouse, |ui| {
-                            ui.checkbox(&mut self.static_enabled, "Use static position");
-                        });
-                        if !self.follow_mouse && self.static_enabled {
-                            ui.horizontal(|ui| {
-                                ui.label("X");
-                                ui.add(egui::DragValue::new(&mut self.static_x));
-                                ui.label("Y");
-                                ui.add(egui::DragValue::new(&mut self.static_y));
-                                ui.label("W");
-                                ui.add(egui::DragValue::new(&mut self.static_w));
-                                ui.label("H");
-                                ui.add(egui::DragValue::new(&mut self.static_h));
-                                if ui.button("Snapshot").clicked() {
-                                    self.static_x = app.window_pos.0;
-                                    self.static_y = app.window_pos.1;
-                                    self.static_w = app.window_size.0;
-                                    self.static_h = app.window_size.1;
-                                }
-                            });
-                        }
-
-                        ui.separator();
-                        ui.heading("Dashboard");
-                        ui.checkbox(
-                            &mut self.dashboard_enabled,
-                            "Enable dashboard when query is empty",
-                        );
-                        ui.horizontal(|ui| {
-                            ui.label("Dashboard config path");
-                            ui.text_edit_singleline(&mut self.dashboard_path);
-                        });
-                        ui.horizontal(|ui| {
-                            ui.label("Default location");
-                            ui.text_edit_singleline(&mut self.dashboard_default_location);
-                        });
-                        ui.checkbox(
-                            &mut self.dashboard_show_when_empty,
-                            "Show dashboard when the search box is blank",
-                        );
-                        ui.checkbox(
-                            &mut self.reduce_dashboard_work_when_unfocused,
-                            "Reduce dashboard work when not focused",
-                        );
-                        if self.debug_logging || cfg!(debug_assertions) {
-                            ui.checkbox(
-                                &mut self.show_dashboard_diagnostics,
-                                "Show dashboard diagnostics (dev)",
-                            );
-                        }
-                        if ui.button("Customize Dashboard...").clicked() {
-                            app.show_dashboard_editor = true;
-                        }
-
-                        ui.separator();
-                        if ui
-                            .button(if self.plugins_expanded {
-                                "Collapse plugin sections"
-                            } else {
-                                "Expand plugin sections"
-                            })
-                            .clicked()
-                        {
-                            self.plugins_expanded = !self.plugins_expanded;
-                            self.expand_request = Some(self.plugins_expanded);
-                        }
-                        let enabled_list = app.enabled_plugins_list();
-                        let mut open_mg_settings_dialog = false;
-
-                        for plugin in app.plugins.iter_mut() {
-                            let name = plugin.name().to_string();
-                            if name == "notes" {
-                                continue;
-                            }
-                            let enabled = match &enabled_list {
-                                Some(list) => list.contains(&name),
-                                None => true,
-                            };
-                            if !enabled {
-                                continue;
-                            }
-
-                            let has_settings = plugin.default_settings().is_some()
-                                || self.plugin_settings.contains_key(&name);
-                            if !has_settings {
-                                continue;
-                            }
-
-                            let entry =
-                                self.plugin_settings.entry(name.clone()).or_insert_with(|| {
-                                    plugin.default_settings().unwrap_or(serde_json::Value::Null)
-                                });
-                            let id = ui.make_persistent_id(format!("plugin_{name}"));
-                            let mut state =
-                                egui::collapsing_header::CollapsingState::load_with_default_open(
-                                    ui.ctx(),
-                                    id,
-                                    false,
-                                );
-                            if let Some(open) = self.expand_request {
-                                state.set_open(open);
-                            }
-                            state
-                                .show_header(ui, |ui| {
-                                    ui.label(format!("{name} settings"));
-                                })
-                                .body(|ui| {
-                                    if name == "mouse_gestures" {
-                                        ui.label("Mouse gesture settings are managed in a dedicated dialog.");
-                                        ui.add_space(6.0);
-                                        if ui.button("Open Mouse Gesture Settings...").clicked() {
-                                            open_mg_settings_dialog = true;
-                                        }
-                                        ui.add_space(4.0);
-                                        ui.small("Tip: you can also open this window via `mg settings`.");
-                                    } else {
-                                        plugin.settings_ui(ui, entry);
-                                    }
-                                });
-                        }
-
-                        if open_mg_settings_dialog {
-                            app.open_mouse_gesture_settings_dialog();
-                        }
-                        let id = ui.make_persistent_id("plugin_notes");
-                        let mut state =
-                            egui::collapsing_header::CollapsingState::load_with_default_open(
-                                ui.ctx(),
-                                id,
-                                false,
-                            );
-                        if let Some(open) = self.expand_request {
-                            state.set_open(open);
-                        }
-                        state
-                            .show_header(ui, |ui| {
-                                ui.label("Note settings");
-                            })
-                            .body(|ui| {
-                                ui.checkbox(
-                                    &mut self.note_save_on_close,
-                                    "Save note on close (Esc)",
-                                );
-                                ui.checkbox(
-                                    &mut self.note_always_overwrite,
-                                    "Always overwrite existing notes",
-                                );
-                                ui.checkbox(
-                                    &mut self.note_images_as_links,
-                                    "Display images as links",
-                                );
-                                ui.checkbox(
-                                    &mut self.note_show_details,
-                                    "Show note details by default",
-                                );
-                                ui.horizontal(|ui| {
-                                    ui.label("Tag/link preview limit");
-                                    ui.add(
-                                        egui::DragValue::new(&mut self.note_more_limit)
-                                            .clamp_range(1..=usize::MAX),
-                                    );
-                                });
-                                let mut cfg = self
-                                    .plugin_settings
-                                    .get("note")
-                                    .and_then(|v| {
-                                        serde_json::from_value::<NotePluginSettings>(v.clone()).ok()
-                                    })
-                                    .unwrap_or_default();
-                                egui::ComboBox::from_label("Open externally")
-                                    .selected_text(match cfg.external_open {
-                                        NoteExternalOpen::Neither => "Neither",
-                                        NoteExternalOpen::Powershell => "Powershell",
-                                        NoteExternalOpen::Notepad => "Notepad",
-                                        NoteExternalOpen::Wezterm => "WezTerm",
-                                    })
-                                    .show_ui(ui, |ui| {
-                                        ui.selectable_value(
-                                            &mut cfg.external_open,
-                                            NoteExternalOpen::Neither,
-                                            "Neither",
-                                        );
-                                        ui.selectable_value(
-                                            &mut cfg.external_open,
-                                            NoteExternalOpen::Powershell,
-                                            "Powershell",
-                                        );
-                                        ui.selectable_value(
-                                            &mut cfg.external_open,
-                                            NoteExternalOpen::Notepad,
-                                            "Notepad",
-                                        );
-                                        ui.selectable_value(
-                                            &mut cfg.external_open,
-                                            NoteExternalOpen::Wezterm,
-                                            "WezTerm",
-                                        );
-                                    });
-                                self.plugin_settings.insert(
-                                    "note".into(),
-                                    serde_json::to_value(cfg).unwrap_or(serde_json::Value::Null),
-                                );
-                            });
-
-                        self.expand_request = None;
-                    });
-
-                ui.separator();
-                ui.horizontal(|ui| {
-                    if ui.button("Save").clicked() {
-                        self.save_settings(ctx, app);
-                    }
-                });
-            });
-        app.show_settings = open;
-    }
-
-    fn save_settings(&mut self, ctx: &egui::Context, app: &mut LauncherApp) {
-        if parse_hotkey(&self.hotkey).is_none() {
-            self.hotkey = self.last_valid_hotkey.clone();
-            self.hotkey_valid = true;
-            app.add_error_toast("Failed to save settings: hotkey is invalid");
-        } else if self.quit_hotkey_enabled && parse_hotkey(&self.quit_hotkey).is_none() {
-            self.quit_hotkey = self.last_valid_quit_hotkey.clone();
-            self.quit_hotkey_valid = true;
-            app.add_error_toast("Failed to save settings: quit hotkey is invalid");
-        } else if self.help_hotkey_enabled && parse_hotkey(&self.help_hotkey).is_none() {
-            self.help_hotkey = self.last_valid_help_hotkey.clone();
-            self.help_hotkey_valid = true;
-            app.add_error_toast("Failed to save settings: help hotkey is invalid");
-        } else {
-            self.last_valid_hotkey = self.hotkey.clone();
-            if self.quit_hotkey_enabled {
-                self.last_valid_quit_hotkey = self.quit_hotkey.clone();
-            }
-            if self.help_hotkey_enabled {
-                self.last_valid_help_hotkey = self.help_hotkey.clone();
-            }
-            self.sync_from_plugin_settings();
-            match Settings::load(&app.settings_path) {
-                Ok(current) => {
-                    let new_settings = self.to_settings(&current);
-                    if let Err(e) = new_settings.save(&app.settings_path) {
-                        app.report_ui_error(UiErrorEvent::new(
-                            "settings_editor.save",
-                            format!("Failed to save: {e}"),
-                        ));
-                    } else {
-                        app.update_paths(
-                            new_settings.plugin_dirs.clone(),
-                            new_settings.index_paths.clone(),
-                            new_settings.enabled_plugins.clone(),
-                            new_settings.enabled_capabilities.clone(),
-                            new_settings.offscreen_pos,
-                            Some(new_settings.enable_toasts),
-                            Some(new_settings.show_inline_errors),
-                            Some(new_settings.show_error_toasts),
-                            Some(new_settings.toast_duration),
-                            Some(new_settings.fuzzy_weight),
-                            Some(new_settings.usage_weight),
-                            Some(new_settings.match_exact),
-                            Some(new_settings.follow_mouse),
-                            Some(new_settings.static_location_enabled),
-                            new_settings.static_pos,
-                            new_settings.static_size,
-                            Some(new_settings.hide_after_run),
-                            Some(new_settings.clear_query_after_run),
-                            Some(new_settings.require_confirm_destructive),
-                            Some(new_settings.timer_refresh),
-                            Some(new_settings.disable_timer_updates),
-                            Some(new_settings.preserve_command),
-                            Some(new_settings.query_autocomplete),
-                            Some(new_settings.net_refresh),
-                            Some(new_settings.net_unit),
-                            new_settings.screenshot_dir.clone(),
-                            Some(new_settings.screenshot_save_file),
-                            Some(new_settings.screenshot_use_editor),
-                            Some(new_settings.screenshot_auto_save),
-                            Some(new_settings.always_on_top),
-                            Some(new_settings.page_jump),
-                            Some(new_settings.note_panel_default_size),
-                            Some(new_settings.note_save_on_close),
-                            Some(new_settings.note_always_overwrite),
-                            Some(new_settings.note_images_as_links),
-                            Some(new_settings.note_show_details),
-                            Some(new_settings.note_more_limit),
-                            Some(new_settings.show_dashboard_diagnostics),
-                        );
-                        ctx.send_viewport_cmd(egui::ViewportCommand::WindowLevel(
-                            if new_settings.always_on_top {
-                                egui::WindowLevel::AlwaysOnTop
-                            } else {
-                                egui::WindowLevel::Normal
-                            },
-                        ));
-                        app.hotkey_str = new_settings.hotkey.clone();
-                        app.quit_hotkey_str = new_settings.quit_hotkey.clone();
-                        app.help_hotkey_str = new_settings.help_hotkey.clone();
-                        app.query_scale = new_settings.query_scale.unwrap_or(1.0).min(5.0);
-                        app.list_scale = new_settings.list_scale.unwrap_or(1.0).min(5.0);
-                        app.history_limit = new_settings.history_limit;
-                        app.clipboard_limit = new_settings.clipboard_limit;
-                        app.page_jump = new_settings.page_jump;
-                        app.note_show_details = new_settings.note_show_details;
-                        app.preserve_command = new_settings.preserve_command;
-                        app.clear_query_after_run = new_settings.clear_query_after_run;
-                        app.require_confirm_destructive = new_settings.require_confirm_destructive;
-                        app.query_autocomplete = new_settings.query_autocomplete;
-                        app.query_results_layout = new_settings.query_results_layout.clone();
-                        app.recompute_query_results_layout();
-                        app.net_refresh = new_settings.net_refresh;
-                        app.net_unit = new_settings.net_unit;
-                        app.screenshot_dir = new_settings.screenshot_dir.clone();
-                        app.screenshot_save_file = new_settings.screenshot_save_file;
-                        app.screenshot_auto_save = new_settings.screenshot_auto_save;
-                        app.screenshot_use_editor = new_settings.screenshot_use_editor;
-                        app.reduce_dashboard_work_when_unfocused =
-                            new_settings.reduce_dashboard_work_when_unfocused;
-                        app.show_dashboard_diagnostics = new_settings.show_dashboard_diagnostics;
-                        app.dashboard_enabled = new_settings.dashboard.enabled;
-                        app.dashboard_show_when_empty =
-                            new_settings.dashboard.show_when_query_empty;
-                        app.dashboard_default_location =
-                            new_settings.dashboard.default_location.clone();
-                        app.dashboard_path = DashboardConfig::path_for(
-                            new_settings
-                                .dashboard
-                                .config_path
-                                .as_deref()
-                                .unwrap_or("dashboard.json"),
-                        )
-                        .to_string_lossy()
-                        .to_string();
-                        app.dashboard.set_path(&app.dashboard_path);
-                        app.toast_duration = new_settings.toast_duration;
-                        app.note_more_limit = new_settings.note_more_limit;
-                        let dirs = new_settings.plugin_dirs.clone().unwrap_or_default();
-                        let actions_arc = Arc::clone(&app.actions);
-                        app.plugins.reload_from_dirs(
-                            &dirs,
-                            app.clipboard_limit,
-                            app.net_unit,
-                            false,
-                            &new_settings.plugin_settings,
-                            actions_arc,
-                        );
-                        if let Some(val) = new_settings.plugin_settings.get("note") {
-                            if let Ok(cfg) =
-                                serde_json::from_value::<NotePluginSettings>(val.clone())
-                            {
-                                app.note_external_open = cfg.external_open;
-                            }
-                        }
-                        crate::request_hotkey_restart(new_settings);
-                        if app.enable_toasts {
-                            app.add_toast(Toast {
-                                text: "Settings saved".into(),
-                                kind: ToastKind::Success,
-                                options: ToastOptions::default()
-                                    .duration_in_seconds(app.toast_duration as f64),
-                            });
-                        }
-                    }
-                }
-                Err(e) => {
-                    let msg = format!("Failed to read settings: {e}");
-                    app.report_error_message("settings_editor.read", msg);
-                }
-            }
-        }
-    }
-}
+pub use state::SettingsEditor;
 
 #[cfg(test)]
 mod tests {
@@ -1066,44 +78,5 @@ mod tests {
         assert!(!saved.static_location_enabled);
         assert_eq!(saved.static_pos, None);
         assert_eq!(saved.static_size, None);
-    }
-
-    #[test]
-    fn query_results_layout_round_trip_editor_conversion() {
-        let mut initial = Settings::default();
-        initial.query_results_layout.enabled = true;
-        initial.query_results_layout.rows = 6;
-        initial.query_results_layout.cols = 4;
-        initial.query_results_layout.respect_plugin_capability = false;
-        initial.query_results_layout.plugin_opt_out = vec!["note".into(), "todo".into()];
-
-        let editor = SettingsEditor::new(&initial);
-        let saved = editor.to_settings(&initial);
-        assert_eq!(saved.query_results_layout, initial.query_results_layout);
-    }
-
-    #[test]
-    fn query_results_layout_clamps_rows_and_cols_to_one() {
-        let initial = Settings::default();
-        let mut editor = SettingsEditor::new(&initial);
-        editor.query_results_layout_enabled = true;
-        editor.query_results_layout_rows = 0;
-        editor.query_results_layout_cols = 0;
-
-        let saved = editor.to_settings(&initial);
-        assert_eq!(saved.query_results_layout.rows, 1);
-        assert_eq!(saved.query_results_layout.cols, 1);
-    }
-
-    #[test]
-    fn error_visibility_round_trip_editor_conversion() {
-        let mut initial = Settings::default();
-        initial.show_inline_errors = false;
-        initial.show_error_toasts = false;
-
-        let editor = SettingsEditor::new(&initial);
-        let saved = editor.to_settings(&initial);
-        assert!(!saved.show_inline_errors);
-        assert!(!saved.show_error_toasts);
     }
 }
